[SpringBoot Service][Ray] refactor: add existid checking in delete by id

+++ src/main/java/com/thoughtworks/springbootemployee/service/CompanyService.java
@@ -46,7 +46,9 @@ public class CompanyService {
     }
 
     public void deleteCompanyById(String companyId) {
-        companyRepository.deleteById(companyId);
+        if (companyRepository.existsById(companyId)){
+            companyRepository.deleteById(companyId);
+        }
     }
 
     public List<Employee> getEmployeesByCompanyId(String companyId) throws EmployeeNotFoundException, CompanyNotFoundException {
